requirements.txt
python-telegram-bot==21.10
APScheduler==3.11.0
pytz==2025.1
python-dotenv==1.1.0
